AuthUser/requirements.txt
Flask
requests
datetime
python_json_config
pyjwt
flask-sqlalchemy
pyodbc